structuratie van parsers & wijzigen models voor parsers
