+++ src/model/Station.java
@@ -1,5 +1,7 @@
 package model;
 
+import java.util.List;
+
 /**
  * Created by Manish on 29/10/2016.
  */
@@ -7,6 +9,7 @@ public class Station {
     private int stationId;
     private String naam;
     private String stad;
+    private List<Trein> treinen;
 
     public Station(int stationId, String naam, String stad) {
         this.stationId = stationId;
@@ -39,5 +42,12 @@ public class Station {
 
     public void setNaam(String naam) {
         this.naam = naam;
+    }
+
+    public List<Trein> getTreinen() {
+        return treinen;
+    }
+    public void setTreinen(List<Trein> treinen) {
+        this.treinen = treinen;
     }
 }
